.hackathon/shipcall/requirements.txt
patter
fastapi
uvicorn[standard]
python-dotenv
httpx
